fix(widgets): apply sync linkages in widget snapshots
